Move ProductActionType and ProductStateType to productTypes file
